refactor(USR-03): dto 디렉터리 수정, 응답 dto 구조 개선
