+++ src/main/java/com/advanceschedular/controller/MemberController.java
@@ -1,16 +1,6 @@
 package com.advanceschedular.controller;
 
-import com.advanceschedular.dto.MemberSignInRequest;
-import com.advanceschedular.dto.MemberSignUpRequest;
-import com.advanceschedular.model.Member;
-import com.advanceschedular.service.MemberService;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpSession;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
