+++ src/main/java/com/advanceschedular/service/MemberService.java
@@ -3,8 +3,8 @@ package com.advanceschedular.service;
 import com.advanceschedular.common.enums.ErrorCode;
 import com.advanceschedular.common.exception.CustomException;
 import com.advanceschedular.common.util.PasswordUtil;
-import com.advanceschedular.dto.MemberSignInRequest;
-import com.advanceschedular.dto.MemberSignUpRequest;
+import com.advanceschedular.dto.member.MemberSignInRequest;
+import com.advanceschedular.dto.member.MemberSignUpRequest;
 import com.advanceschedular.model.Member;
 import com.advanceschedular.repository.MemberRepository;
 import jakarta.transaction.Transactional;
